+++ packages/views/example/index.js
@@ -1,10 +1,14 @@
 import React from 'react';
+import Helmet from 'react-helmet';
 import Container from './Container';
 
 class Example extends React.Component {	
 	render () {
 	    return (
-	        <Container/>		
+	        <div>
+	            <Helmet title="Example page" />
+				<Container/>				
+	        </div>
 	    );
 	}
 }