+++ packages/web/index.jsx
@@ -1,39 +1,9 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import {render} from 'react-dom';
-import {Provider} from 'react-redux';
-import {ConnectedRouter} from 'react-router-redux';
-import {history, store} from '@qasir/store';
-import Routes from '@qasir/routes';
-
-let contentBuffer = {
-    pathName: null,
-    content: null,
-};
-
-class Application extends Component {
-    static childContextTypes = {
-        getContentBuffer: PropTypes.func,
-        setContentBuffer: PropTypes.func,
-    }
-
-    getChildContext() {
-        return {
-            getContentBuffer: () => contentBuffer,
-            setContentBuffer: ({ pathName, content }) => (contentBuffer = { pathName, content }),
-        }
-    }
-
-    shouldComponentUpdate() {
-        return true
-    }
-
-    render() {
-        return (
-            <Routes />
-        )
-    }
-}
+import React from 'react'
+import {render} from 'react-dom'
+import {Provider} from 'react-redux'
+import {ConnectedRouter} from 'react-router-redux'
+import {history, store} from '@qasir/store'
+import Application from './component/application.jsx'
 
 render(
     <Provider store={store}>
@@ -41,4 +11,16 @@ render(
             <Application/>            
         </ConnectedRouter>
     </Provider>, 
-document.getElementById('application'));
+document.getElementById("application"))
+
+if (process.env.PWA) {
+    if ('serviceWorker' in navigator) {    
+        window.addEventListener('load', () => {
+            navigator.serviceWorker.register('/service-worker.js').then(registration => {
+                console.log('SW registered: ', registration);
+            }).catch(registrationError => {
+                console.log('SW registration failed: ', registrationError);
+            });
+        });
+    }
+}