+++ CHANGELOG.md
@@ -10,3 +10,5 @@
     - Update package position
     - Adding hot module reloading
     - Adding ESLint code linting (Airbnb linter)
+    - Update routing system
+    - Adding PWA config support (workbox)